#18 Example music sheet to not auto-load bigdata on page load
