+++ src/app/test/unit/sheet.test.ts
@@ -1,24 +1,4 @@
 describe("Component: Sheet", () => {
-  const data = [
-    "X:1\nK:D\nDD AA|BBA2|\n",
-    "X:1\nT:The Sheet\nM:4/4\nC:Trad.\nK:G\n|:Gccc dedB|dedB dedB|c2ec B2dB|c2A2 A2BA|",
-    `
-      X:1
-      T:alternate heads
-      M:C
-      L:1/8
-      U:n=!style=normal!
-      K:C treble style=rhythm
-      "Am" BBBB B2 B>B | "Dm" B2 B/B/B "C" B4 |"Am" B2 nGnB B2 nGnA | "Dm" nDB/B/ nDB/B/ "C" nCB/B/ nCB/B/ |B8| B0 B0 B0 B0 |]
-      %%text This translates to:
-      [M:C][K:style=normal]
-      [A,EAce][A,EAce][A,EAce][A,EAce] [A,EAce]2 [A,EAce]>[A,EAce] |[DAdf]2 [DAdf]/[DAdf]/[DAdf] [CEGce]4 |[A,EAce]2 GA [A,EAce] GA |D[DAdf]/[DAdf]/ D[DAdf]/[DAdf]/ C [CEGce]/[CEGce]/ C[CEGce]/[CEGce]/ |[CEGce]8 | [CEGce]2 [CEGce]2 [CEGce]2 [CEGce]2 |]
-      GAB2 !style=harmonic![gb]4|GAB2 [K: style=harmonic]gbgb|
-      [K: style=x]
-      C/A,/ C/C/E C/zz2|
-      w:Rock-y did-nt like that`,
-  ];
-
   it("should verify that testing works!", () => {
     expect(true).toBe(true);
   });
